cypress only: disable most l10n tasks
As it turns out, disabling l10n is not a well supported thing, especially for Nightly builds.

This disables everything except the initial repacks on Linux. These are tricky to remove, but cheap and only run on the Nightly graph anyways, so it's not worth the effort of hacking them out.